worker/requirements.txt
pika==1.3.2
psycopg2-binary==2.9.9
requests==2.32.3
beautifulsoup4==4.12.3
